# docs/swagger.yaml
basePath: /
definitions:
  controllers.ChangePasswordRequest:
    properties:
      password_baru:
        type: string
      password_lama:
        type: string
    type: object
...
  controllers.UpdateProfileRequest:
    properties:
      email:
        type: string
      nama:
        type: string
    type: object
...
info:
  contact: {}
  description: Dokumentasi API backend Sistem Informasi Karang Taruna
  title: API Sistem Informasi Karang Taruna
  version: "1.0"
paths:
...
  /api/profile:
    get:
      description: Mengambil profile user yang sedang login
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get profile
      tags:
      - Profile
    put:
      consumes:
      - application/json
      description: Mengubah nama dan email user yang sedang login
      parameters:
      - description: Data update profile
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update profile
      tags:
      - Profile
  /api/profile/foto:
    post:
      consumes:
      - multipart/form-data
      description: Upload foto profile user yang sedang login
      parameters:
      - description: File foto profile
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Upload foto profile
      tags:
      - Profile
  /api/profile/password:
    put:
      consumes:
      - application/json
      description: Mengubah password user yang sedang login
      parameters:
      - description: Data ubah password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Change password
      tags:
      - Profile
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"